Add a flag to make a client wait for an empty line before exiting in control mode to avoid stray commands ending up in the shell.

+++ server.c
@@ -45,7 +45,7 @@ struct clients		 clients;
 
 struct tmuxproc		*server_proc;
 static int		 server_fd = -1;
-static int		 server_client_flags;
+static uint64_t		 server_client_flags;
 static int		 server_exit;
 static struct event	 server_ev_accept;
